requirements.txt
msgpack~=1.0.4
requests~=2.27.1
Flask==2.1.3
PyYAML~=6.0
UnityPy~=1.9.4
asyncio~=3.4.3
Pillow~=9.1.0
waitress~=2.1.1
python-dateutil~=2.8.2
pydantic~=1.9.0
colorama~=0.4.4
websocket-client~=1.3.2
numpy~=1.22.3
pytz~=2022.1
aiofiles~=0.8.0
wordcloud~=1.8.2.2
jieba~=0.42.1
matplotlib~=3.5.2
configobj~=5.0.6
APScheduler~=3.9.1
html2text~=2020.1.16
xmltodict~=0.13.0
mutagen~=1.45.1
pydub~=0.25.1